Kubernetes Validating Admission Controller 101 project
This validating admission controller watches (only) the creation
of Pods and disallows to create Pods named pod-with-an-invalid-name.
All other Pods names are allowed.

It is inspired by this project and maintains the same Apache License 2.0:
https://github.com/stackrox/admission-controller-webhook-demo

See more information below on the why section and the differences between the projects.

Target audience:
* Developers getting started with implementing Validating Admission Controllers in Kubernetes
* Developers keen to quickly debug and isolate issues related to Validating Admission Controllers
* Curious souls